inserindo campo last_update

+++ src/main/java/com/natal/subscriptionservice/infrastructure/repository/ClientRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClientRepository extends JpaRepository<ClientEntity, Integer> {
+public interface ClientRepository extends JpaRepository<ClientEntity, Long> {
     ClientEntity findByDocument(String document);
 }
